requirements.txt
discord.py
python-dotenv
PyNaCl>=1.5.0
